+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **[FIX]** Fix back button alignment in `SearchForm` autocomplete sections.
 - **[UPDATE]** Only show the `SearchForm` switch button when at least FROM or TO is selected.
 [...]
+- **[FIX]** Remove hover color & pointer when `disabled` is passed to `ItemCheckbox`/`ItemRadio`
 
 # v34.1.0 (27/05/2020)